mini_project/pytweet/requirements.txt
Flask==0.10.1
SQLAlchemy==0.8.3
psycopg2==2.5.1
behave==1.2.2
nose==1.3.0
pyhamcrest==1.8.0
Flask-Testing==0.4